chore: move checkStaticPage into config
