+++ src/config/check-static-page.ts
